Move 'profile.js' to 'components/profile' folder
