+++ client/src/app.js
@@ -11,7 +11,7 @@ import Map from './pages/map';
 import NoMatch from './pages/no-match';
 import Events from './pages/events';
 import Restaurant from './components/restaurant/restaurant-item';
-import Profile from './components/profile';
+import Profile from './components/profile/profile';
 
 class App extends Component {
   render() {
